requirements.txt
# Discord Bot Framework
discord.py>=2.3.2

# Claude AI Integration
anthropic>=0.39.0

# RCON Client for Project Zomboid
mcrcon>=0.7.0

# Environment Variables
python-dotenv>=1.0.0

# Async HTTP (for Steam Workshop API)
aiohttp>=3.9.0

# HTML Parsing (for mod info scraping)
beautifulsoup4>=4.12.0

# Task Scheduling
APScheduler>=3.10.0

# Logging
colorlog>=6.8.0

# RAG Knowledge Base
chromadb>=0.4.22
sentence-transformers>=2.3.0
requests>=2.31.0
